requirements.txt
# ML core
torch>=2.1,<2.10
torchvision>=0.16,<0.25
segmentation-models-pytorch>=0.3.3

# Data & math
numpy>=1.24,<1.27
pandas>=1.5,<2.2
scipy>=1.9,<1.14

# Image & GIS
pillow>=9.0,<11.0
rasterio>=1.3,<1.4
shapely>=2.0,<2.1
geopandas>=0.12,<0.15
scikit-image>=0.19,<0.23

# Web frameworks
fastapi>=0.95,<0.115
uvicorn[standard]>=0.21,<0.32
streamlit>=1.21,<1.40

# Utils
requests>=2.28,<2.33
aiofiles>=23.0,<24.2
python-multipart>=0.0.6,<0.0.10

# Visualization
matplotlib>=3.5,<3.10
seaborn>=0.12,<0.14
plotly>=5.13,<5.25



